rustc: middle: move some types from resolve to privacy.
